提供Pojav后端支持

+++ LWJGL-H2CO3/src/main/java/org/lwjgl/glfw/GLFWDropCallback.java
@@ -5,15 +5,13 @@
  */
 package org.lwjgl.glfw;
 
-import static org.lwjgl.glfw.GLFW.glfwSetDropCallback;
-import static org.lwjgl.system.MemoryUtil.NULL;
-import static org.lwjgl.system.MemoryUtil.memGetAddress;
-import static org.lwjgl.system.MemoryUtil.memUTF8;
+import javax.annotation.*;
 
-import org.lwjgl.system.Callback;
-import org.lwjgl.system.Pointer;
+import org.lwjgl.system.*;
 
-import javax.annotation.Nullable;
+import static org.lwjgl.system.MemoryUtil.*;
+
+import static org.lwjgl.glfw.GLFW.*;
 
 /**
  * Instances of this class may be passed to the {@link GLFW#glfwSetDropCallback SetDropCallback} method.
@@ -75,7 +73,7 @@ public abstract class GLFWDropCallback extends Callback implements GLFWDropCallb
      * @return the name at the specified index as a String
      */
     public static String getName(long names, int index) {
-        return memUTF8(memGetAddress(names + (long) Pointer.POINTER_SIZE * index));
+        return memUTF8(memGetAddress(names + Pointer.POINTER_SIZE * index));
     }
 
     /** See {@link GLFW#glfwSetDropCallback SetDropCallback}. */
